==> Code/1- Getting Number of Buses.ipynb <==
import pandas as pd
import numpy as np

# import the trips csv as a dataframe 
df = pd.read_csv("trips.csv")

# remove unnessecary rows
df.drop(["route_desc", "route_long", "route_type", "service_id", "route_colo", "route_text", 
         "wheelchair", "UID", "O_AREA", "D_AREA", "TRIP_ASOD", "VEH_ID", "AC", "DIGI_DATE", "DIGI_SPE", 
         "DIGI_NOTES", "ASSIGN_ID", "H3RES5_IDX", "RANKING", "INST_NUM", "FARE_Q219"], axis=1, inplace= True)
df.head()

### Subset Data - Keep only Cairo Transport Authority (CTA) routes

bool_series = df['agency_id'] == 'CTA'  #true for all rows that have agency_id = CTA
df_CTA = df[bool_series]

# sort by route number (to check if trips in same route have same headway : TRUE)
# reset index with drop = true so that a new column isn't added: https://stackoverflow.com/questions/16167829/in-pandas-how-can-i-reset-index-without-adding-a-new-column
df_CTA = df_CTA.sort_values('route_shor', ascending= 1).reset_index(drop=True)


### Get number of Buses: 

#### 1) trip duration

# get number of buses
# assume a constant speed of 15km/h across network (4.2m/s)
# multiply by 2 to get the time for the whole route
df_CTA['route_time_s'] = ((df_CTA["LEN_KM"] *1000) / 4.2) * 2
df_CTA.head()

#### 2) no. of buses

# get no. of buses for each trip (route time/ headway)/2
# divide by 2 because this is the number of buses for the whole route (the rows are trips: each route has two trips)
df_CTA['buses'] = ((df_CTA["route_time_s"]/df_CTA["headway_se"]) / 2)
df_CTA.head()

# get sum of buses to compare to reported total buses of 2650
df_CTA["buses"].sum()

### get ratio of calculated sum to reported sum

total_buses = 2650
ratio = total_buses/df_CTA["buses"].sum()   # 18.69


### Multiply bus column by ratio to reach figure closer to 2650

# .apply(np.ceil) to round values up (we don't want any trips with 0 values)
df_CTA['buses'] = (df_CTA['buses']*ratio).apply(np.ceil)
##df_CTA.head()

# we need the number of buses for each route to be split equally between trips
# the number of buses for each trip was calculated seperately based on the trip length (varies between outbound and return)
# this groups by route, and assigns both trips the lowest of the two 'buses' value calculated
# https://stackoverflow.com/questions/49591867/create-new-column-with-max-value-with-groupby
df_CTA['standard_buses'] = df_CTA.groupby(['route_id'])['buses'].transform(min)
df_CTA.head()

### see what the sum is now: it needs to be close to 2650

# see what the sum is now
df_CTA["standard_buses"].sum()

df_CTA["standard_buses"].describe()

### save file to csv

df_CTA.to_csv(r'buses_per_trip.csv', index = None, header=True)
