==> Code/2- Distance Matrix.ipynb <==
import pandas as pd
import numpy as np

# import the distance matrix csv as a dataframe 
df = pd.read_csv("terminal-garage-distance-v2.csv")

df.head()

### Multiply Distance by 2 to account for Pull in and Pull out

df.Distance = df.Distance * 2
df

# Pivot DataFrame so that the garages are columns
# https://stackoverflow.com/questions/44513488/how-to-remove-multilevel-index-in-pandas-pivot-table
df = df.pivot(index='InputID', columns='TargetID', values = 'Distance').reset_index().rename_axis(None, axis=1)
df.head()

# keep only the garages used by the CTA buses. There are four garages that are used by CTA minibuses (these are dropped)
df.drop(['Katameya', 'Al Mataria', 'El Basatin', 'Tiba'], axis = 1, inplace = True)
df.head()

print(df.loc[0, 'InputID'])

#### save file to csv

df.to_csv(r'distance_matrix.csv', index = None, header=True)
