==> Code/3- Append Distance Matrix to Trips.ipynb <==
import pandas as pd
import numpy as np

#### import both csvs 

df_trips = pd.read_csv("buses_per_trip.csv")
df_trips.head()

df_distances = pd.read_csv("distance_matrix.csv")
df_distances.head()

## Merge Dataframes so that you know the distance from the start point of each trip to all depots

# how = 'left' joins over the left dataframe, in this case we are joining onto the df_trips dataframe
df_bus_to_garage = pd.merge(df_trips, df_distances, left_on='O_TERMINAL', right_on='InputID', how='left')
df_bus_to_garage.head()

df_bus_to_garage.to_csv(r'bus-to-garage.csv', index = None, header=True)
