==> Code/4- Optimization.ipynb <==
import numpy as np
import pandas as pd

### CVXPY used to formulate optimization problem
### GUROBI is a powerful commercial solver that is free for academics

import cvxpy as cp
from gurobipy import *

load in the csvs that will be used in the optimization problem

df_distance_matrix = pd.read_csv("bus-to-garage.csv")

# remove unnessecary rows
df_distance_matrix.drop(["route_id", "route_shor", "trip_id", "trip_heads", "trip_short", "direction_", "agency_id", 
         "headway_se", "O_TERMINAL", "LEN_KM", "STD_FARE", "route_time_s", "buses", "standard_buses", "InputID"], axis=1, inplace= True)
df_distance_matrix.head()

Transform dataframe to a numpy array

np_distance_matrix = df_distance_matrix.to_numpy()
#show the array
np_distance_matrix

check dimensions

np.shape(np_distance_matrix)

The depots are in alphabetical order. The csv with depot capacities needs to have the depots in alphabetical order as well!

IMPORTING CSV WITH GARAGE CAPACITIES

df_capacities = pd.read_csv("Garage Capacities.csv")
list(df_capacities.columns) 

df_capacities.drop(["Name"], axis=1, inplace= True)
df_capacities.head()

convert to a numpy array

np_depot_capacities = df_capacities.to_numpy()
# show the output
np_depot_capacities

np.shape(np_depot_capacities)

We need to change this so that the capacities are all in one row. This way it will match the distance matrix array and
we can sum over the columns

# use flatten to collapse array into one dimension
np_depot_capacities = np_depot_capacities.flatten()
np_depot_capacities

We need another array for the number of buses on each trip. This is necessary for the depot capacity constraint

# read in the csv
df_buses = pd.read_csv("bus-to-garage.csv")
# keep only the buses column
df_buses = df_buses[["standard_buses"]]
df_buses.head()

np_buses = df_buses.to_numpy()
np_buses = np_buses.flatten()
np_buses

# we need to repeat each row 20 times, where 20 is the number of depots
# np.tile 'tiles' the array 20 times (20 rows of 340 values)
# .T transposes the array so that it is 340 rows with each row having the number of buses 
# of the respective trip repeated 20 times
buses = np.tile(np_buses,(20,1)).T
buses
#np.shape(buses)    # (340, 20)

# SETTING UP THE PROBLEM

## Problem is set up as an LP Relaxation Problem. This is because it cannot be solved as an IP (340*20 variables = 6800 : time taken is 2^3600)
## Branch and Bound is then used to find approximate integer values

### selection is a binary '0/1' matrix. 1 means bus is assigned to terminal. If binary then (I think) the problem is np complete and solver keeps running for hours, an alternative is to allow positive fractions

selection = cp.Variable(shape=buses.shape,boolean=True)
#try without boolean
#selection = cp.Variable(shape=buses.shape, nonneg=True)



## Constraints

### trip can only be assigned to one depot (Row sums need to equal 1)

assignment_constraint = cp.sum(selection,axis=1) == 1

### depots cannot exceed capacities. Column sums <= depot capacity

#### cp.multiply(selection, buses) gets the number of buses that will be assigned to a garage. We are treating this as an all or nothing assignment problem. This means that either all buses from the same route are assigned to a garage or non of them are (they are not split up between garages)

capacity_constraint = cp.sum(cp.multiply(selection, buses), axis=0) <= np_depot_capacities

#### Add a max distance constraint 
##### 20000 equivalent to 20km

distance_constraint = cp.sum(cp.multiply(selection, np_distance_matrix), axis=1) <= 50000

### Minimize dead mileage. Dead mileage is the product of the distance matrix and the (buses*selection)

cost = cp.sum(cp.multiply(cp.multiply(selection, buses), np_distance_matrix))

Check installed solvers

print (cp.installed_solvers())

### Run the optimization problem. MAJOR FLAW (SOME ROWS ARE NOT ASSIGNED AS ALL COLUMNS FOR THAT ROW ARE <0.5 AND SO ROUND TO 0. IS THIS A FLAW IN ECOS_BB?

constraints = [assignment_constraint, capacity_constraint]

assign_problem = cp.Problem(cp.Minimize(cost),constraints)


assign_problem.solve(solver= cp.GUROBI, verbose = True)

# use ECOS_BB for branch and bound (should return integer solutions: some rows sum to zero after rounding :( )
#assign_problem.solve(solver= cp.ECOS_BB)

# ECOS_BB with max iterations: used when selection variable is boolean (IP problem). Doesn't find a solution
#assign_problem.solve(solver= cp.ECOS_BB, mi_max_iters= 1000)



print(selection.value)
print('Problem status: {}'.format(assign_problem.status))

Now save the output as a csv. We can use this to check that the assignment_constraint was not violated

# save before rounding 
np.savetxt("assignment.csv", selection.value, delimiter=",")

type(selection.value)

## No need for this. GUROBI SOLVES IT USING BRANCH AND BOUND

# round numpy array
#selection.value = selection.value.round()
#np.savetxt("assignment.csv", selection.value, delimiter=",")

Let's check if buses were allocated correctly. This shows how many buses were allocated to each garage. We use this to check that the garage capacities were not exceeded (capacity_constraint not violated)

allocation = selection.value*buses
np.savetxt("allocation.csv", allocation, delimiter=",")

### Get dead mileage per trip

test = (selection.value*buses) * np_distance_matrix

np.savetxt("total_mileage_per_trip.csv", test, delimiter=",")

#### Dead Mileage per trip (not weighted by number of buses on each trip)

test2 = (selection.value * np_distance_matrix)

np.savetxt("mileage_per_trip.csv", test2, delimiter=",")
