==> Code/5- Plotting Results.ipynb <==
import numpy as np
import pandas as pd

# read in results. header = None used so that first row is not treated as a header
df_results = pd.read_csv("allocation.csv", header = None)
df_results.head()

## Get the names of the garages in a list to append as a header to the results

garages_df = pd.read_csv("distance_matrix.csv")
# remove 1st column which is the terminal associated with the garage
del garages_df['InputID']
garages_df.head()


## Extract header as list

garages = list(garages_df.columns.values)
garages

## Add list as headers using .columns

df_results.columns = garages
df_results.head()

df_results.sum(axis = 0) 

df_summary = df_results.sum(axis = 0) 
type(df_summary)

### we need to convert it to a dataframe

df_summary = df_summary.to_frame()

### Reset index and add column names

# reset index so that Terminal Name is a Column, not the index
df_summary.reset_index(inplace=True)
#Add Column headers (to merge with later)
#headers = ['Name', 'Buses Allocates']
df_summary.columns = ['Name', 'Buses_Allocated']
df_summary

### Add the Garage Capacities csv to merge them together

#### .str.strip() removes whitespace. Otherwise we are unable to merge

df_capacities = pd.read_csv("Garage Capacities.csv")
# strip function - removes all whitespace from all stringlike cells in dataframe
# https://stackoverflow.com/questions/33788913/pythonic-efficient-way-to-strip-whitespace-from-every-pandas-data-frame-cell-tha
df_capacities = df_capacities.apply(lambda x: x.str.strip() if x.dtype == "object" else x)

### Merge

# how = 'left' joins over the left dataframe, in this case we are joining onto the df_trips dataframe
df_plot = pd.merge(df_summary, df_capacities, left_on='Name', right_on='Name', how='left')
df_plot

#df_plot.to_csv('Assignment Plots\\garage_capacity.csv')

df_plot.to_csv(r'/Users/Hussein/Documents/Masters/UCL/Classes/Term 1/Quantitative Methods/Written Investigation/Code/Assignment Plots/df_plot.csv')

## MORE PREP FOR R PLOTS

#### Average Dead Mileage Per Trip

# read in results. header = None used so that first row is not treated as a header
df_deadMile = pd.read_csv("mileage_per_trip.csv", header = None)
df_deadMile.head()

#### Convert to Numpy array to get the maximum value from each row (the only non-zero value)

np_deadMile = df_deadMile.values
np_deadMile

#### Get dead mileage for each trip

deadMileAll = np.amax(np_deadMile, axis=1)   # Maxima along the second axis (rows)

df_deadMilePlot = pd.DataFrame(deadMileAll)
df_deadMilePlot 

# Get number of buses on each trip - to append to the dataframe above. This way we can plot a histogram for the dead mileage of all buses, not just all trips

# read in the csv
df_buses = pd.read_csv("bus-to-garage.csv")
# keep only the buses column
df_buses = df_buses[["standard_buses"]]
df_buses

# Merge by index

merged = pd.merge(df_deadMilePlot,df_buses,left_index=True,right_index=True)
merged

#### save to plot in R

merged.to_csv(r'/Users/Hussein/Documents/Masters/UCL/Classes/Term 1/Quantitative Methods/Written Investigation/Code/Assignment Plots/df_deadMilePlot.csv')

### Repeat each row based on number of buses value (to plot histogram on bus level not trip level)

merged = merged.reindex(merged.index.repeat(merged.standard_buses))

merged.to_csv(r'/Users/Hussein/Documents/Masters/UCL/Classes/Term 1/Quantitative Methods/Written Investigation/Code/Assignment Plots/df_deadMilePlotRep.csv')
